compat: remove warning for < 4.1
It still is sort of experimental, I suppose, especially this part in the
udp_tunnel drop-in:

	skb_orphan(skb);
	sk_mem_reclaim(sk);

It seems like sometimes this won't do what we want, but it's hard to
diagnose exactly what's happening. In any case, nobody paid attention to
that warning anyway, so let's just get rid of it.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>