drivers: rmnet: perf: Automatic flush after chain
For newer targets which employ dl marker v2 logic,
we are able to blindly flush at the end of skb
chains since the chains are up to 20x times
longer than past targets. To differentiate
between targets we check to see if dl marker
v2 is in use during flush decisions at the
end of a chain.

Change-Id: I0a8f4745dd6dfba5ce5585010d7e5cae3cf4f486
Signed-off-by: Conner Huff <chuff@codeaurora.org>