Kbuild: remove -fvisibility=hidden from cflags
This was originally done in 2015 as a means of decreasing module size,
but it has the effect of creating JUMP11 relocations on ARM when
compiled in THUMB2 mode without CONFIG_THUMB2_AVOID_R_ARM_THM_JUMP11=y,
which results in `B ...` instructions being generated with jumps that
are too far, rather than `B.W ...` instructions, which can handle the
larger sized jump.

Get rid of the old hack, which had minimum utility anyway.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>