Revert "compat: skb_mark_not_on_list will be backported to Ubuntu 18.04"
This reverts commit 25c887b725e5687fb6fd87c0ea7b7fce173c0d89.

Because this commit has not been backported so far, due to the implications
of building Ubuntu's backport of wireguard in a timely manner.

For now, reverting this fix would allow wireguard-linux-compat CI to work
on Ubuntu 18.04.

A different fix or the same one can be applied again when the time is
right.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>