extcon: Allocate blocking_notifier_head after registering the device
Commit a137d00ed85aea5ecc10b2d2c5bc48d4268a11cf
("extcon: Modify extcon device to be created after driver data is set")
move device_register right before adding extcon_dev_list.

This is fine on upstream, however since we have blocking notifier and
it uses managed kzalloc (which has to be used when the dev is already
registered successfully), we have to move it as well.

Fixes: dfcf2d9f2ee3f99d0131b0f6a5deefcd0f2c1ae8
("extcon: Add usage of blocking notifier chain")
Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>
Signed-off-by: azrim <mirzaspc@gmail.com>